添加了CFG的漏洞查找
1.未完成数据库是否启动部分的检查代码
2.界面部分没问题,后台处理代码还未完成
3.计算的数据存入了数据库,并且有查重处理

+++ code-similarity-site/mysite/settings.py
@@ -172,6 +172,6 @@ def make_base_dirs():
     if not os.path.isdir(NEO4J_DATABASE_PATH):
         os.makedirs(NEO4J_DATABASE_PATH)
         os.makedirs(os.path.join(NEO4J_DATABASE_PATH, "vuln_db", "create_logs"))
-        if os.path.isdir(NEO4J_HOME):
-            shutil.copytree(NEO4J_HOME, os.path.join(NEO4J_DATABASE_PATH, "neo4j"))
+    if os.path.isdir(NEO4J_HOME) and not os.path.exists(os.path.join(NEO4J_DATABASE_PATH, "neo4j")):
+        shutil.copytree(NEO4J_HOME, os.path.join(NEO4J_DATABASE_PATH, "neo4j"))
